# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/client/search": {
      "get": {
        "operationId": "ClientController_search",
        "summary": "Search user data",
        "parameters": [
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "description": "Client ID of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dbcoll",
            "required": false,
            "in": "query",
            "description": "Database collection name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelLink",
            "required": false,
            "in": "query",
            "description": "Channel link of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "link",
            "required": false,
            "in": "query",
            "description": "Link of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Name of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "required": false,
            "in": "query",
            "description": "Phone number of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "required": false,
            "in": "query",
            "description": "Password of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repl",
            "required": false,
            "in": "query",
            "description": "Repl link of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientName",
            "required": false,
            "in": "query",
            "description": "Clientname of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deployKey",
            "required": false,
            "in": "query",
            "description": "Deployment key URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mainAccount",
            "required": false,
            "in": "query",
            "description": "Main account of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product",
            "required": false,
            "in": "query",
            "description": "Product associated with the client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the searched user data."
          }
        },
        "tags": [
          "Clients"
        ]
      }
    },
    "/client/{clientId}": {
      "get": {
        "operationId": "ClientController_findOne",
        "summary": "Get user data by ID",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the user data."
          },
          "404": {
            "description": "User data not found."
          }
        },
        "tags": [
          "Clients"
        ]
      },
      "patch": {
        "operationId": "ClientController_update",
        "summary": "Update user data by ID",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user data has been successfully updated."
          },
          "404": {
            "description": "User data not found."
          }
        },
        "tags": [
          "Clients"
        ]
      },
      "delete": {
        "operationId": "ClientController_remove",
        "summary": "Delete user data by ID",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user data has been successfully deleted."
          },
          "404": {
            "description": "User data not found."
          }
        },
        "tags": [
          "Clients"
        ]
      }
    },
    "/client/query": {
      "post": {
        "operationId": "ClientController_executeQuery",
        "summary": "Execute a custom MongoDB query",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Query executed successfully."
          },
          "400": {
            "description": "Invalid query."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "Clients"
        ]
      }
    },
...
    "/bufferclient": {
      "post": {
        "operationId": "BufferClientController_create",
        "summary": "Create user data",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBufferClientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user data has been successfully created."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "BufferClients"
        ]
      },
      "get": {
        "operationId": "BufferClientController_findAll",
        "summary": "Get all user data",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all user data."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "BufferClients"
        ]
      }
    },
    "/bufferclient/search": {
      "get": {
        "operationId": "BufferClientController_search",
        "summary": "Search user data",
        "parameters": [
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "description": "Client ID of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dbcoll",
            "required": false,
            "in": "query",
            "description": "Database collection name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelLink",
            "required": false,
            "in": "query",
            "description": "Channel link of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "link",
            "required": false,
            "in": "query",
            "description": "Link of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Name of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "required": false,
            "in": "query",
            "description": "Phone number of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "required": false,
            "in": "query",
            "description": "Password of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repl",
            "required": false,
            "in": "query",
            "description": "Repl link of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientName",
            "required": false,
            "in": "query",
            "description": "Clientname of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deployKey",
            "required": false,
            "in": "query",
            "description": "Deployment key URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mainAccount",
            "required": false,
            "in": "query",
            "description": "Main account of the client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product",
            "required": false,
            "in": "query",
            "description": "Product associated with the client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the searched user data."
          }
        },
        "tags": [
          "BufferClients"
        ]
      }
    },
    "/bufferclient/{bufferClientId}": {
      "get": {
        "operationId": "BufferClientController_findOne",
        "summary": "Get user data by ID",
        "parameters": [
          {
            "name": "bufferClientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the user data."
          },
          "404": {
            "description": "User data not found."
          }
        },
        "tags": [
          "BufferClients"
        ]
      },
      "patch": {
        "operationId": "BufferClientController_update",
        "summary": "Update user data by ID",
        "parameters": [
          {
            "name": "bufferClientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user data has been successfully updated."
          },
          "404": {
            "description": "User data not found."
          }
        },
        "tags": [
          "BufferClients"
        ]
      },
      "delete": {
        "operationId": "BufferClientController_remove",
        "summary": "Delete user data by ID",
        "parameters": [
          {
            "name": "bufferClientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user data has been successfully deleted."
          },
          "404": {
            "description": "User data not found."
          }
        },
        "tags": [
          "BufferClients"
        ]
      }
    },
    "/bufferclient/query": {
      "post": {
        "operationId": "BufferClientController_executeQuery",
        "summary": "Execute a custom MongoDB query",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Query executed successfully."
          },
          "400": {
            "description": "Invalid query."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "BufferClients"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS and Express API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateBufferClientDto": {
        "type": "object",
        "properties": {
          "channelLink": {
            "type": "string",
            "example": "paid_giirl_shruthiee",
            "description": "Channel link of the user"
          },
          "dbcoll": {
            "type": "string",
            "example": "shruthi",
            "description": "Database collection name"
          },
          "link": {
            "type": "string",
            "example": "PaidGirl.netlify.app/Shruthi1",
            "description": "Link of the user"
          },
          "name": {
            "type": "string",
            "example": "Shruthi Reddy",
            "description": "Name of the user"
          },
          "number": {
            "type": "string",
            "example": "+916265240911",
            "description": "Phone number of the user"
          },
          "password": {
            "type": "string",
            "example": "Ajtdmwajt1@",
            "description": "Password of the user"
          },
          "repl": {
            "type": "string",
            "example": "https://shruthi1.glitch.me",
            "description": "Repl link of the user"
          },
          "session": {
            "type": "string",
            "example": "1BQANOTEuMTA4LjUg==",
            "description": "Session token"
          },
          "userName": {
            "type": "string",
            "example": "ShruthiRedd2",
            "description": "Username of the user"
          },
          "clientId": {
            "type": "string",
            "example": "shruthi1",
            "description": "Client ID of the user"
          },
          "deployKey": {
            "type": "string",
            "example": "https://shruthi1.glitch.me/exit",
            "description": "Deployment key URL"
          },
          "mainAccount": {
            "type": "string",
            "example": "ShruthiRedd2",
            "description": "Main account of the user"
          },
          "product": {
            "type": "string",
            "example": "booklet_10",
            "description": "Product associated with the user"
          }
        },
        "required": [
          "channelLink",
          "dbcoll",
          "link",
          "name",
          "number",
          "password",
          "repl",
          "session",
          "userName",
          "clientId",
          "deployKey",
          "mainAccount",
          "product"
        ]
      },
...
    }
  }
}